% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fft.R
\name{compute_fft}
\alias{compute_fft}
\title{compute_fft()}
\usage{
compute_fft(incoming)
}
\arguments{
\item{incoming}{data.frame containing the incoming waveform for the FFT.}
}
\value{
A data.frame with the following columns: .values with the values from the FFT, .idx with the indexes for these values (this makes it easier to plot with ggplot later). .psd contains the power spectral density of the fourier coefficient.
}
\description{
Computes the FFT of the given waveform.
}
\examples{
waveform(duration_s = 1.0, sr = 10) \%>\%
  cos_sum(freqs = c(2.0, 3.0), amplitudes = c(0.5, 1.0)) \%>\%
  length_norm() \%>\%
  compute_fft()
}
\references{
Brunton, S. L. & Kutz, J. N. Data-Driven Science and Engineering: Machine Learning, Dynamical Systems, and Control. (Cambridge University Press, 2022). doi:10.1017/9781009089517. Pg 68.
}
