% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_decay.R
\name{exp_decay}
\alias{exp_decay}
\title{exp_decay()}
\usage{
exp_decay(incoming, decay_tau)
}
\arguments{
\item{incoming}{Required. A data.frame of the incoming waveform.}

\item{decay_tau}{Required. Rate of the decay.}
}
\value{
Returns a data.frame in the format of the waveform() function describing the product of the incoming waveform with the exponential decay specified by this function.
}
\description{
Multiplies incoming waveform by an exponential decay curve.
}
\examples{
waveform() \%>\%
  cos_sum(freqs = c(1, 3), amplitudes = NULL, phases = NULL) \%>\%
  exp_decay(decay_tau = 0.5)
}
