% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_plot.R
\name{fft_plot}
\alias{fft_plot}
\title{fft_plot()}
\usage{
fft_plot(incoming, show = "psd", ...)
}
\arguments{
\item{incoming}{Required. A data.frame with the incoming waveform.}

\item{show}{Optional. Either "psd" (which shows the power spectral density of the FFT up to half of its values to display frequencies up to the Nyquist limit) or "everything" (which shows modulus, real, and imaginary components of all FFT values.)}

\item{...}{Optional. If specified, passed to the ggplot2 labs() function.}
}
\value{
ggplot() object of the FFT
}
\description{
Plots the results of an FFT. It converts the x-axis scale to hz during the plotting.
}
\examples{
waveform(duration_s = 1.0, sr = 10) \%>\%
  cos_sum(freqs = c(2.0, 3.0), amplitudes = c(0.5, 1.0)) \%>\%
  length_norm() \%>\%
  compute_fft() \%>\%
  fft_plot(show = "everything")

waveform(duration_s = 1.0, sr = 10) \%>\%
 cos_sum(freqs = c(2.0, 3.0), amplitudes = c(0.5, 1.0)) \%>\%
 length_norm() \%>\%
 compute_fft() \%>\%
 fft_plot(show = "psd")
}
