% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_fft.R
\name{inverse_fft}
\alias{inverse_fft}
\title{inverse_fft()}
\usage{
inverse_fft(incoming_waveform, incoming_fft)
}
\arguments{
\item{incoming_waveform}{Required. A data.frame specifying an incoming waveform. Hint: this incoming waveform can be created by the waveform() function.}

\item{incoming_fft}{Required. The incoming FFT as computed by compute_fft()}
}
\value{
A data.frame with the result of the inverse FFT.
}
\description{
Runs the inverse FFT specified by incoming_fft, applies the result to incoming_waveform, and returns the result. Unlilke the functions that sum the incoming waveform, this function overwrites the incoming waveform.
}
\examples{
wv <- waveform(duration_s = 1.0, sr = 200)
wv_fft <- wv \%>\%
  cos_sum(freqs = c(2.0, 3.0), amplitudes = c(0.5, 1.0)) \%>\%
  length_norm() \%>\%
  compute_fft()
wv \%>\%
  inverse_fft(wv_fft) \%>\%
  waveform_plot()
}
