% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_norm.R
\name{length_norm}
\alias{length_norm}
\title{length_norm()}
\usage{
length_norm(incoming)
}
\arguments{
\item{incoming}{Required. A data.frame with the waveform.}
}
\value{
A data.frame with the normalized waveform.
}
\description{
Normalizes a waveform by dividing the vector of values by the length of the vector of values. Used to prepare waveforms for FFT.
}
\examples{
waveform(duration_s = 1.0, sr = 100) \%>\%
  cos_sum(freqs = c(1.0, 3.0), amplitudes = NULL, phases = NULL) \%>\%
  length_norm()
}
