% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveform.R
\name{waveform}
\alias{waveform}
\title{waveform
Builds the skeleton of a waveform for a given duration and sample rate filled with zero values.}
\usage{
waveform(duration_s = 1, sr = 200)
}
\arguments{
\item{duration_s}{Waveform duration in seconds.}

\item{sr}{Waveform sample rate in samples / second.}
}
\value{
Returns a data.frame with the following columns: .sample, the integer sample of the row; .sec, the double second of that row; and .value, the value for that row.
}
\description{
waveform
Builds the skeleton of a waveform for a given duration and sample rate filled with zero values.
}
\examples{
waveform(duration_s = 1.0, sr = 200)
}
