% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveform_plot.R
\name{waveform_plot}
\alias{waveform_plot}
\title{waveform_plot()}
\usage{
waveform_plot(incoming, ...)
}
\arguments{
\item{incoming}{Required. A data.frame with the incoming waveform to plot.}

\item{...}{Optional. If specified, extra arguments are passed to the ggplot2 labs() function to place labels on the plot.}
}
\value{
Returns a ggplot object that can be saved or displayed.
}
\description{
waveform_plot()
}
\examples{
waveform() \%>\% cos_sum(1) \%>\% waveform_plot()
}
