% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_noise.R
\name{white_noise}
\alias{white_noise}
\title{white_noise()}
\usage{
white_noise(incoming, mean = 0, sd = 1)
}
\arguments{
\item{incoming}{Required. The incoming waveform to which Gaussian will be added to.}

\item{mean}{Optional. The mean of the Gaussian distribution of the noise. Defaults to 0.}

\item{sd}{Optional. The standard deviation of the Gaussian distribution of the noise. Defaults to 1.}
}
\value{
data.frame of the waveform with noise added.
}
\description{
Adds Gaussian white noise into a waveform. You should set the random seed before invoking this function if you want consistent results.
}
\examples{
waveform() \%>\% white_noise() \%>\% waveform_plot()
}
